+++ Cogs/Utils/__init__.py
@@ -1,0 +1,4 @@
+from .ping import Ping_Meta
+
+async def setup(bot):
+    await bot.add_cog(Ping_Meta(bot))