+++ main.py
@@ -1,24 +1,37 @@
-import discord
-from dotenv import load_dotenv
-import os
+from colorama import Fore, Style, init
 from discord.ext import commands
+from dotenv import load_dotenv
+from Configs import bot_config
+import os
+import discord
 
-#* Loading .env variables
+# Loading .env variables
 load_dotenv()
 TOKEN = os.getenv("TOKEN")
+PREFIX = bot_config["General"]["Prefix"]
 
-#* Configuring intents
+# Configuring intents
 intents = discord.Intents.default()
 intents.message_content = True
 
-#* Declaring MyClient
-MyClient = commands.Bot(command_prefix="^",intents=intents)
+# Declaring MyClient
+class MyBot(commands.Bot):
+    async def setup_hook(self):
+        # Loading da cogs
+        cogs = ["Utils cogs"]
+        #// await self.load_extension("Cogs.Moderation")
+        #// await self.load_extension("Cogs.Fun")
+        await self.load_extension("Cogs.Utils")
+        for cog in cogs:
+            print(f"{cog:<18}[{Fore.GREEN}OK{Style.RESET_ALL}]")
+MyClient = MyBot(command_prefix="^", intents=intents, help_command=None)
 
-#* On ready event
+# On ready event
 @MyClient.event
 async def on_ready():
-    print(f"{MyClient.user} Serving 🛜...\nZip ID 🪪: {MyClient.user.id}")
-print(f"token: {TOKEN}")
-#* Initializing bot
+    activity = discord.Streaming(name="Fundamental Paper Education",url="https://www.youtube.com/watch?v=0x7zdgK-FFU",platform="YouTube")
+    await MyClient.change_presence(activity=activity)
+    print(f"{MyClient.user} Serving 🛜...\nZip ID 🪪: {MyClient.user.id}\nStreaming success ✅.")
+# Initializing bot
 if __name__ == "__main__":
     MyClient.run(TOKEN)