+++ changelog.md
@@ -1,5 +1,24 @@
 # CHANGELOG
 
+## 1.1.0
+### Added
+- `requeriments.txt` file in root
+- `Cogs/` directory in root
+- `Utils/` directory in `Cogs/`
+- `ping.py` command in `Utils/`
+
+- `Configs/` directory in root
+- `config.toml` file in `Configs/`
+- `toml_loader.py` in `Configs/`
+
+- `Helpers/` directory in root
+- `auto_init_metaclass.py` in `Helpers/`
+
+### Changed
+- Modified `main.py` adding better UX when initializing the bot
+
+---
+
 ## 1.0.0
 ### Added
 - `.gitignore` file